backend/requirements.txt

pillow==5.4.1
django-crispy-forms==1.7.2
django-ckeditor==5.6.1

